requirements.txt
Flask == 2.0.2
requests == 2.26.0
urllib3 == 1.26.7
python-dotenv = 0.19.2
